feat: refactor ButtonScrollToTop component to use custom hook for scroll visibility logic
